backend/pyproject.toml
[tool.poetry]
name = "military-civilian-resume-backend"
version = "0.1.0"
description = "Backend services for translating Canadian Armed Forces MPRR data into civilian resume narratives."
authors = ["Automated Agent <agent@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = { extras = ["standard"], version = "^0.31.1" }
pydantic = "^2.5.0"
pyyaml = "^6.0.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.1.0"
playwright = "^1.40.0"
mcp = "^1.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
httpx = "^0.27.1"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
